+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.22.0]
 
+* Change linters to no longer create and write to an "OUTPUT" pane channel
 * Fix file recompilation to clear previous diagnostics for the file
 * Fix possible JavaScript error in the "Go to Symbol in Editor..." implementation